requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
decorator==4.4.2
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
moviepy==1.0.3
numpy==1.24.4
orjson==3.9.5
packaging==23.1
passlib==1.7.4
Pillow==10.0.0
pluggy==1.3.0
proglog==0.1.10
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.16.2
